Extract common factory behavior
